# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/menus":{"get":{"summary":"Show Menus","operationId":"show_menus_api_v1_menus_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MenuShow"},"type":"array","title":"Response Show Menus Api V1 Menus Get"}}}}}},"post":{"summary":"Create Menu","operationId":"create_menu_api_v1_menus_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuChange"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseMenu"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/menus/{menu_id}":{"get":{"summary":"Get Menu","operationId":"get_menu_api_v1_menus__menu_id__get","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Menu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseMenu"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Update Menu","operationId":"update_menu_api_v1_menus__menu_id__patch","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Menu Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuChange"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseMenu"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Menu","operationId":"delete_menu_api_v1_menus__menu_id__delete","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Menu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/menus/{menu_id}/submenus":{"get":{"summary":"Get Submenus","operationId":"get_submenus_api_v1_menus__menu_id__submenus_get","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Menu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SubMenuShow"},"title":"Response Get Submenus Api V1 Menus  Menu Id  Submenus Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Submenu","operationId":"create_submenu_api_v1_menus__menu_id__submenus_post","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Menu Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubMenuChange"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseSubMenu"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/menus/{menu_id}/submenus/{submenu_id}":{"get":{"summary":"Get Submenu","operationId":"get_submenu_api_v1_menus__menu_id__submenus__submenu_id__get","parameters":[{"name":"submenu_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Submenu Id"}},{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Menu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseSubMenu"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Update Submenu","operationId":"update_submenu_api_v1_menus__menu_id__submenus__submenu_id__patch","parameters":[{"name":"submenu_id","in":"path","required":true,"schema":{"type":"string","title":"Submenu Id"}},{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","title":"Menu Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubMenuChange"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseSubMenu"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Update Submenu","operationId":"update_submenu_api_v1_menus__menu_id__submenus__submenu_id__delete","parameters":[{"name":"menu_id","in":"path","required":true,"schema":{"type":"string","title":"Menu Id"}},{"name":"submenu_id","in":"path","required":true,"schema":{"type":"string","title":"Submenu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/menus/{menu_id}/submenus/{submenu_id}/dishes":{"get":{"summary":"Get Dishes","operationId":"get_dishes_api_v1_menus__menu_id__submenus__submenu_id__dishes_get","parameters":[{"name":"submenu_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Submenu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DishShow"},"title":"Response Get Dishes Api V1 Menus  Menu Id  Submenus  Submenu Id  Dishes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Dish","operationId":"create_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes_post","parameters":[{"name":"submenu_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Submenu Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DishChange"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseDish"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/menus/{menu_id}/submenus/{submenu_id}/dishes/{dish_id}":{"get":{"summary":"Get Dish","operationId":"get_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__get","parameters":[{"name":"dish_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Dish Id"}},{"name":"submenu_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Submenu Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseDish"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Update Dish","operationId":"update_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__patch","parameters":[{"name":"submenu_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Submenu Id"}},{"name":"dish_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Dish Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DishChange"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseDish"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Submenu","operationId":"delete_submenu_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__delete","parameters":[{"name":"submenu_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Submenu Id"}},{"name":"dish_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Dish Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BaseDish":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"price":{"type":"string","title":"Price"}},"type":"object","required":["title","description","price"],"title":"BaseDish"},"BaseMenu":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"}},"type":"object","required":["title","description"],"title":"BaseMenu"},"BaseSubMenu":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"}},"type":"object","required":["title","description"],"title":"BaseSubMenu"},"DishChange":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"price":{"type":"string","title":"Price"}},"type":"object","required":["title","description","price"],"title":"DishChange"},"DishShow":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"price":{"type":"string","title":"Price"}},"type":"object","required":["title","description","price"],"title":"DishShow"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MenuChange":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"}},"type":"object","required":["title","description"],"title":"MenuChange"},"MenuShow":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"submenus_count":{"type":"integer","title":"Submenus Count"},"dishes_count":{"type":"integer","title":"Dishes Count"}},"type":"object","required":["title","description","submenus_count","dishes_count"],"title":"MenuShow"},"SubMenuChange":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"}},"type":"object","required":["title","description"],"title":"SubMenuChange"},"SubMenuShow":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"dishes_count":{"type":"integer","title":"Dishes Count"}},"type":"object","required":["title","description","dishes_count"],"title":"SubMenuShow"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}